# backend/turismo-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Turismo Necochea",
        "description": "Documentación de la API para la aplicación de Turismo de Necochea",
        "contact": {
            "email": "tudai@alumnos.exa.unicen.edu.ar"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Servidor principal de la API"
        }
    ],
    "paths": {
...
        "/api/complejos": {
            "get": {
                "tags": [
                    "Complejos"
                ],
                "summary": "Obtener todos los complejos",
                "operationId": "31f0d6a7ad46a3451c44e14391416e90",
                "responses": {
                    "200": {
                        "description": "Lista de complejos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Complejo"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "Complejos"
                ],
                "summary": "Crear un nuevo complejo",
                "operationId": "a3686b102094dfa5233c141d6a4e7b19",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ComplejoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Complejo creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Complejo"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/complejos/{id}": {
            "get": {
                "tags": [
                    "Complejos"
                ],
                "summary": "Obtener un complejo por ID",
                "operationId": "40574db9d42ea05e60f119050bc845b3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del complejo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del complejo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Complejo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Complejo no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Complejos"
                ],
                "summary": "Actualizar un complejo por ID",
                "operationId": "886a2ba4a90b92e0deebee038a009f27",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del complejo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ComplejoUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Complejo actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Complejo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Complejo no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Complejos"
                ],
                "summary": "Eliminar un complejo por ID",
                "operationId": "e6325034bc8e898e48081d678158604b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del complejo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Complejo eliminado correctamente"
                    },
                    "404": {
                        "description": "Complejo no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Complejo": {
                "title": "Complejo",
                "description": "Modelo de un complejo",
                "properties": {
                    "id": {
                        "description": "ID único del complejo",
                        "type": "integer",
                        "format": "int64"
                    },
                    "nombre": {
                        "description": "Nombre del complejo",
                        "type": "string"
                    },
                    "direccion": {
                        "description": "Dirección del complejo",
                        "type": "string"
                    },
                    "mail": {
                        "description": "Correo electrónico de contacto",
                        "type": "string",
                        "format": "email",
                        "nullable": true
                    },
                    "redesSociales": {
                        "description": "URL de redes sociales",
                        "type": "string",
                        "nullable": true
                    },
                    "telefono": {
                        "description": "Número de teléfono de contacto",
                        "type": "string",
                        "nullable": true
                    },
                    "servicio": {
                        "description": "Descripción de los servicios ofrecidos",
                        "type": "string",
                        "nullable": true
                    },
                    "adicional": {
                        "description": "Información adicional del complejo",
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ComplejoRequest": {
                "title": "ComplejoRequest",
                "description": "Esquema para la creación de un complejo",
                "required": [
                    "nombre",
                    "direccion"
                ],
                "properties": {
                    "nombre": {
                        "type": "string",
                        "example": "Complejo Las Dunas"
                    },
                    "direccion": {
                        "type": "string",
                        "example": "Av. Costanera 1000"
                    },
                    "mail": {
                        "type": "string",
                        "format": "email",
                        "example": "info@lasdunas.com",
                        "nullable": true
                    },
                    "redesSociales": {
                        "type": "string",
                        "example": "http://instagram.com/lasdunas",
                        "nullable": true
                    },
                    "telefono": {
                        "type": "string",
                        "example": "2262123456",
                        "nullable": true
                    },
                    "servicio": {
                        "type": "string",
                        "example": "Canchas de tenis, piscina, restaurante",
                        "nullable": true
                    },
                    "adicional": {
                        "type": "string",
                        "example": "Abierto todo el año",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ComplejoUpdateRequest": {
                "title": "ComplejoUpdateRequest",
                "description": "Esquema para la actualización de un complejo",
                "properties": {
                    "nombre": {
                        "type": "string",
                        "example": "Complejo El Faro",
                        "nullable": true
                    },
                    "direccion": {
                        "type": "string",
                        "example": "Calle 4 N° 200",
                        "nullable": true
                    },
                    "mail": {
                        "type": "string",
                        "format": "email",
                        "example": "contacto@elfaro.com",
                        "nullable": true
                    },
                    "redesSociales": {
                        "type": "string",
                        "example": "http://facebook.com/elfaro",
                        "nullable": true
                    },
                    "telefono": {
                        "type": "string",
                        "example": "2262987654",
                        "nullable": true
                    },
                    "servicio": {
                        "type": "string",
                        "example": "Spa, gimnasio",
                        "nullable": true
                    },
                    "adicional": {
                        "type": "string",
                        "example": "Eventos corporativos",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}